requirements.txt
discord.py ~= 1.7.3
mcstatus ~= 9.0.4
jishaku ~= 2.5.0